requirements.txt
telethon>=1.36.0
pybit>=5.6.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiosqlite>=0.20.0
